+++ bootstrap.sh
@@ -10,6 +10,8 @@ if [ -x "$(which python)" ]; then
     ./py2/bin/pip install pyramid==1.9.4
     ./py2/bin/pip install repoze.zcml==0.4
     ./py2/bin/pip install repoze.workflow==0.6.1
+    ./py2/bin/pip install https://github.com/conestack/odict/archive/master.zip
+    ./py2/bin/pip install https://github.com/conestack/node/archive/master.zip
     ./py2/bin/pip install https://github.com/conestack/node.ext.zodb/archive/master.zip
     ./py2/bin/pip install https://github.com/conestack/yafowil/archive/master.zip
     ./py2/bin/pip install https://github.com/conestack/cone.tile/archive/master.zip
@@ -26,6 +28,8 @@ if [ -x "$(which python3)" ]; then
     ./py3/bin/pip install pyramid==1.9.4
     ./py3/bin/pip install repoze.zcml==1.0b1
     ./py3/bin/pip install repoze.workflow==1.0b1
+    ./py3/bin/pip install https://github.com/conestack/odict/archive/master.zip
+    ./py3/bin/pip install https://github.com/conestack/node/archive/master.zip
     ./py3/bin/pip install https://github.com/conestack/node.ext.zodb/archive/master.zip
     ./py3/bin/pip install https://github.com/conestack/yafowil/archive/master.zip
     ./py3/bin/pip install https://github.com/conestack/cone.tile/archive/master.zip